requirements.txt
asgiref==3.4.1
attrs==20.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==3.2.6
django-colorfield==0.4.2
django-crispy-forms==1.12.0
gunicorn==21.2.0
idna==3.3
jellyfish==0.9.0
Jinja2==3.1.2
lib50==3.0.10
MarkupSafe==2.1.3
packaging==23.2
pexpect==4.8.0
Pillow==10.1.0
ptyprocess==0.7.0
pycparser==2.21
pytz==2021.1
PyYAML==6.0.1
requests==2.27.1
setuptools==69.0.2
sqlparse==0.4.1
submit50==3.1.1
termcolor==1.1.0
typing_extensions==4.2.0
urllib3==1.26.18
wheel==0.42.0
